[FIX] Fix lint and build errors
